+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Fixed issue with passing user variable to Slack integration when an error is triggered
+- Added ```--no-check``` swith to override active file and server checks
 - Upon exit, deploy will now return the repo's current active branch to its original state, instead of assuming checkout of master
 - Improved dependency checks
 - Fixed output from ```deploy --function-list``` and ```deploy --variable-list [project]```